Still updating subtree extraction
